Refactored code. Added interceptors.
