+++ apache-http-client/src/main/java/io/refactoring/http5/client/example/helper/UserSimpleHttpRequestHelper.java
@@ -1,4 +1,4 @@
-package io.refactoring.http5.client.example.util;
+package io.refactoring.http5.client.example.helper;
 
 import io.refactoring.http5.client.example.RequestProcessingException;
 import java.net.URI;
